Feature/ae 1880 php85 compatibility

+++ Classes/PHPExcel/Calculation/Functions.php
@@ -428,7 +428,7 @@ class PHPExcel_Calculation_Functions
             return self::VALUE();
         }
 
-        return ($value % 2 == 0);
+        return ((int) $value % 2 == 0);
     }
 
 
@@ -448,7 +448,7 @@ class PHPExcel_Calculation_Functions
             return self::VALUE();
         }
 
-        return (abs($value) % 2 == 1);
+        return ((int) abs($value) % 2 == 1);
     }
 
 
@@ -547,7 +547,7 @@ class PHPExcel_Calculation_Functions
             case 'integer':
                 return $value;
             case 'boolean':
-                return (integer) $value;
+                return (int) $value;
             case 'string':
                 //    Errors
                 if ((strlen($value) > 0) && ($value[0] == '#')) {
